+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 - Initial release
 - Add Caching strategies;
 - Add Logging strategies;
+- Add Custom parse for JSON responses;
+- Add Response and Exception processors;
+- Add Base client class;
+- Add Basic configuration;